added 5-more_numbers.c

+++ more_functions_nested_loops/5-more_numbers.c
@@ -2,13 +2,13 @@
 #include "main.h"
 
 /**
-* more_numbers -  Prints 10 times the numbers, from 0 to 14
+* more_numbers - Prints 10 times the numbers, from 0 to 14
 * followed by a new line
 */
 
 void more_numbers(void)
 {
-	int row, num;
+	long int row, num;
 
 	for (row = 0; row < 10; row++)
 	{
